Source Codes/metricsManagement/metricsManagement/metricsManagementRepo/metrics_management/requirements.txt
flask==2.3.0
mysql-connector-python-rf
requests